bug(TCOMP-2452) separate HealthcheckServices file
