+++ sample-parent/sample-connector/src/main/resources/org/talend/sdk/component/test/connectors/Messages_zh_CN.properties
